+++ .gitignore
@@ -3,4 +3,4 @@ data/
 .ipynb_checkpoints/
 __pycache__/
 .env
-notebooks/
+tests/
