--- instruments/__init__.py
@@ -1,9 +1,0 @@
-"""
-This __init__.py makes the importing of all Instrument classes more accessible.
-"""
-from .instrument import Instrument, MetaInstrument, PhysicalInstrument
-from .vaunix.labbrick import LabBrick
-from .signal_hound.sa124 import Sa124
-from .meta.cqed_components import QuantumElement, QuantumDevice
-from .meta.stage import Stage
-from .quantum_machines import qm_config_builder
